+++ src/features/survey/Intro.tsx
@@ -1,4 +1,3 @@
-import { useEffect } from 'react';
 import Image from 'next/image';
 import { css, type Theme } from '@emotion/react';
 import { AnimatePresence, m } from 'framer-motion';
@@ -6,13 +5,13 @@ import introBgPng from 'public/images/intro/intro_bg.png';
 import introBgWebp from 'public/images/intro/intro_bg.webp';
 
 import CTAButton from '~/components/button/CTAButton';
+import { useCTAButtonVisible, useParagraphStep } from '~/components/intro/hooks';
 import IntroHeader from '~/components/intro/IntroHeader';
-import StaggerWrapper from '~/components/stagger/StaggerWrapper';
+import SkipStaggerWrapper from '~/components/intro/SkipStaggerWrapper';
 import WatsonCharacter from '~/components/watson/WatsonCharacter';
+import { defaultFadeInVariants } from '~/constants/motions';
 import { RESEARCHER_NAME } from '~/constants/name';
 import { CTAVariants } from '~/features/survey/styles';
-import useBoolean from '~/hooks/common/useBoolean';
-import useStep from '~/hooks/step/useStep';
 
 interface StepProps {
   next?: () => void;
@@ -20,32 +19,36 @@ interface StepProps {
 }
 
 const Intro = ({ next }: StepProps) => {
-  const { currentStep } = useParagraphStep();
+  const { currentStep, paragraphStep } = useParagraphStep();
   const { isCTAButtonVisible } = useCTAButtonVisible();
 
   if (currentStep === 4) {
     return (
-      <section css={sectionCss}>
+      <m.section css={sectionCss} variants={defaultFadeInVariants} initial="initial" animate="animate" exit="exit">
         <picture css={pictureCss}>
           <source srcSet="/images/survey/intro-image-4.webp" type="image/webp" />
           <Image src="/images/survey/intro-image-4.png" alt="나의 질문 폼 생성" fill />
         </picture>
 
         <IntroHeader />
+
         <article css={articleCss}>
-          <AnimatePresence mode="wait">{currentStep === 4 && <Paragraph4 key="4" />}</AnimatePresence>
+          <AnimatePresence mode="wait">
+            {currentStep === 4 && <Paragraph4 key="4" onSkip={paragraphStep} />}
+          </AnimatePresence>
         </article>
+
         {isCTAButtonVisible && (
           <m.div css={fixedBottomCss} variants={CTAVariants} initial="initial" animate="animate" exit="exit">
             <CTAButton onClick={next}>생성하기</CTAButton>
           </m.div>
         )}
-      </section>
+      </m.section>
     );
   }
 
   return (
-    <section css={sectionCss}>
+    <m.section css={sectionCss} variants={defaultFadeInVariants} initial="initial" animate="animate" exit="exit">
       <picture css={pictureCss}>
         <source srcSet={introBgWebp.src} type="image/webp" />
         <Image src={introBgPng} alt="nalab intro" fill />
@@ -56,16 +59,16 @@ const Intro = ({ next }: StepProps) => {
         <AnimatePresence mode="wait">
           {[1, 2, 3].includes(currentStep) && (
             <>
-              {currentStep === 1 && <Paragraph1 key="1" />}
-              {currentStep === 2 && <Paragraph2 key="2" />}
-              {currentStep === 3 && <Paragraph3 key="3" />}
+              {currentStep === 1 && <Paragraph1 key="1" onSkip={paragraphStep} />}
+              {currentStep === 2 && <Paragraph2 key="2" onSkip={paragraphStep} />}
+              {currentStep === 3 && <Paragraph3 key="3" onSkip={paragraphStep} />}
             </>
           )}
         </AnimatePresence>
       </article>
 
       <WatsonCharacter />
-    </section>
+    </m.section>
   );
 };
 
@@ -73,53 +76,55 @@ export default Intro;
 
 const articleCss = css`
   margin-bottom: 42px;
-  padding-top: 126px;
+  padding-top: calc(126px - 56px);
 `;
 
-const Paragraph1 = () => {
+type SkipProps = { onSkip: () => void };
+
+const Paragraph1 = ({ onSkip }: SkipProps) => {
   return (
-    <StaggerWrapper>
+    <SkipStaggerWrapper onSkip={onSkip}>
       <p>안녕하세요!</p>
       <p>저는 당신의 커리어 연구원</p>
       <p>
         <strong>{RESEARCHER_NAME}</strong>이에요.
       </p>
-    </StaggerWrapper>
+    </SkipStaggerWrapper>
   );
 };
 
-const Paragraph2 = () => {
+const Paragraph2 = ({ onSkip }: SkipProps) => {
   return (
-    <StaggerWrapper>
+    <SkipStaggerWrapper onSkip={onSkip}>
       <p>동료의 익명 피드백을 수집해서</p>
       <p>
         당신의 커리어 <strong>DNA</strong>를 찾는
       </p>
       <p>연구를 한답니다.</p>
-    </StaggerWrapper>
+    </SkipStaggerWrapper>
   );
 };
 
-const Paragraph3 = () => {
+const Paragraph3 = ({ onSkip }: SkipProps) => {
   return (
-    <StaggerWrapper>
+    <SkipStaggerWrapper onSkip={onSkip}>
       <p>동료의 익명 피드백은</p>
       <p>
         <strong>나의 질문 폼</strong>을 만들고 공유해서
       </p>
       <p>수집할 수 있어요.</p>
-    </StaggerWrapper>
+    </SkipStaggerWrapper>
   );
 };
 
-const Paragraph4 = () => {
+const Paragraph4 = ({ onSkip }: SkipProps) => {
   return (
-    <StaggerWrapper>
+    <SkipStaggerWrapper onSkip={onSkip}>
       <p>지금, 피드백을 받을 수 있는 </p>
       <p>
         <strong>나의 질문 폼</strong>을 생성해보세요!
       </p>
-    </StaggerWrapper>
+    </SkipStaggerWrapper>
   );
 };
 
@@ -165,36 +170,3 @@ const fixedBottomCss = (theme: Theme) => css`
   max-width: ${theme.size.maxWidth};
   padding: 0 16px;
 `;
-
-const 문구_수 = 4;
-const 매_문구_지속시간 = 3500;
-
-const useParagraphStep = () => {
-  const { currentStep, next: paragraphStep } = useStep({ initial: 1, max: 문구_수 });
-
-  useEffect(() => {
-    const interval = setInterval(paragraphStep, 매_문구_지속시간);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [paragraphStep]);
-
-  return {
-    currentStep,
-  };
-};
-
-const useCTAButtonVisible = () => {
-  const [isCTAButtonVisible, _, setTrue] = useBoolean(false);
-
-  useEffect(() => {
-    const timeout = setTimeout(setTrue, 매_문구_지속시간 * 문구_수);
-
-    return () => {
-      clearTimeout(timeout);
-    };
-  }, [setTrue]);
-
-  return { isCTAButtonVisible };
-};
